Break apart the <Language>Extensions.kt files used in Lambda extension points
