+++ jetbrains-core/resources/META-INF/ext-python.xml
@@ -3,9 +3,9 @@
 
 <idea-plugin>
     <extensions defaultExtensionNs="aws.toolkit.lambda">
-        <runtimeGroup runtimeGroup="PYTHON" implementation="software.aws.toolkits.jetbrains.services.lambda.PythonRuntimeGroup"/>
-        <packager runtimeGroup="PYTHON" implementation="software.aws.toolkits.jetbrains.services.lambda.PythonLambdaPackager"/>
-        <handlerResolver runtimeGroup="PYTHON" implementation="software.aws.toolkits.jetbrains.services.lambda.PythonLambdaHandlerResolver"/>
-        <sam.debugSupport runtimeGroup="PYTHON" implementation="software.aws.toolkits.jetbrains.services.lambda.PythonSamDebugSupport"/>
+        <runtimeGroup runtimeGroup="PYTHON" implementation="software.aws.toolkits.jetbrains.services.lambda.python.PythonRuntimeGroup"/>
+        <packager runtimeGroup="PYTHON" implementation="software.aws.toolkits.jetbrains.services.lambda.python.PythonLambdaPackager"/>
+        <handlerResolver runtimeGroup="PYTHON" implementation="software.aws.toolkits.jetbrains.services.lambda.python.PythonLambdaHandlerResolver"/>
+        <sam.debugSupport runtimeGroup="PYTHON" implementation="software.aws.toolkits.jetbrains.services.lambda.python.PythonSamDebugSupport"/>
     </extensions>
 </idea-plugin>
