+++ jetbrains-core/tst/software/aws/toolkits/jetbrains/services/lambda/python/PythonLambdaHandlerResolverTest.kt
@@ -1,7 +1,7 @@
 // Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.services.lambda
+package software.aws.toolkits.jetbrains.services.lambda.python
 
 import com.intellij.openapi.application.runReadAction
 import com.jetbrains.python.psi.PyFunction
@@ -9,6 +9,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.Rule
 import org.junit.Test
 import software.amazon.awssdk.services.lambda.model.Runtime
+import software.aws.toolkits.jetbrains.services.lambda.Lambda
 import software.aws.toolkits.jetbrains.utils.rules.PythonCodeInsightTestFixtureRule
 
 class PythonLambdaHandlerResolverTest {
@@ -28,7 +29,11 @@ class PythonLambdaHandlerResolverTest {
 
         runReadAction {
             val elements =
-                Lambda.findPsiElementsForHandler(projectRule.project, Runtime.PYTHON3_6, "hello_world.app.handle")
+                Lambda.findPsiElementsForHandler(
+                    projectRule.project,
+                    Runtime.PYTHON3_6,
+                    "hello_world.app.handle"
+                )
             assertThat(elements).hasSize(1)
             assertThat(elements[0]).isInstanceOfSatisfying(PyFunction::class.java) {
                 assertThat(it.qualifiedName).isEqualTo("hello_world.app.handle")
@@ -47,7 +52,11 @@ class PythonLambdaHandlerResolverTest {
         )
 
         runReadAction {
-            val elements = Lambda.findPsiElementsForHandler(projectRule.project, Runtime.PYTHON3_6, "hello_world")
+            val elements = Lambda.findPsiElementsForHandler(
+                projectRule.project,
+                Runtime.PYTHON3_6,
+                "hello_world"
+            )
             assertThat(elements).hasSize(0)
         }
     }
